Separate TimeSeries from line hiding/selection logic for area breakdown
